Home view 1. feat: canales component 2. fix: script attributes 3. feat: Planes Internet from API
